extract getTemplatePackages and template types, gather debug
