+++ packages/geoprocessing/scripts/types.ts
@@ -42,3 +42,10 @@ export interface GeoprocessingBundle extends ProcessingBundle {
 export interface PreprocessingBundle extends ProcessingBundle {
   options: PreprocessingHandlerOptions;
 }
+
+export interface TemplateMetadata {
+  templates: string | string[];
+}
+
+const TemplateTypes = ["add-on-template", "starter-template"] as const;
+export type TemplateType = (typeof TemplateTypes)[number];
